Refactor infobox API and remove memoization
Moved all infobox-related code to a single module, infobox.py. This
simplifies the circular import issues and allows for better reuse of code.

Refactored the Infobox class to have fewer methods with clearer
names. Used properties where possible.

Fixed the attributes() method for an infobox. We now only return
attributes that are actually present in the infobox, instead of
returning all attributes in the meta infobox. Also, made a change to
ignore numeric suffixes when looking up rendered mappings.

Changed infobox attribute names to be dash-separated instead of
underscore_separated. By standardizing on representing attributes
internally in WikipediaBase with dashes, we avoid having to call
replace("-", "_") in multiple places.

Removed the global memoization decorator. This slows things down, but
we're better off without an unbounded cache. It might be worth memoizing
some functions later on. For development, though, removing memoization
exposes the performance bottlenecks with the code (instead of hiding the
issues with a very large cache).
